requirements.txt
asgiref==3.8.1
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.3
django-cors-headers==4.7.0
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.0
djangorestframework-stubs==3.16.5
idna==3.11
iniconfig==2.1.0
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
urllib3==2.5.0
